heatwaveR: Detect Heatwaves and Cold-Spells

The different methods for defining, detecting, and categorising the extreme events known as heatwaves or cold-spells, as first proposed in Hobday et al. (2016) doi: 10.1016/j.pocean.2015.12.014 and Hobday et al. (2018) https://www.jstor.org/stable/26542662. The functions in this package work on both air and water temperature data. These detection algorithms may be used on non-temperature data as well.

See also

Author

Maintainer: Robert W. Schlegel robwschlegel@gmail.com (ORCID) [contributor]

Authors:

Albertus J. Smit (ORCID) [contributor]