Constructs a continuous, uninterrupted time series of temperatures.

Takes a series of dates and temperatures, and if irregular (but ordered), inserts missing dates and fills correpsonding temperatures with NAs.

make_whole(data, x = t, y = temp)

Arguments

data
A data frame with columns for date and temperature data. Ordered daily data are expected, and although missing values (NA) can be accommodated, the function is only recommended when NAs occur infrequently, preferably at no more than 3 consecutive days.

x
A column with the daily time vector (see details). For backwards compatibility, the column is named t by default.

y
A column with the response vector. RmarineHeatWaves version <= 0.15.9 assumed that this would be daily seawater temperatures, but as of version 0.16.0 it may be any arbitrary measurement taken at a daily frequency. The default remains temperature, and the default column name is therefore temp, again hopefully ensuring backwards compatibility.

Value

The function will return a data frame with three columns. The column headed doy (day-of-year) is the Julian day running from 1 to 366, but modified so that the day-of-year series for non-leap-years runs 1...59 and then 61...366. For leap years the 60th day is February 29. See the example, below. The other two columns take the names of x and y, if supplied, or it will be t and temp in case the default values were used. The x (or t) column is a series of dates of class Date, while y (or temp) is the measured variable. This time series will be uninterrupted and continuous daily values between the first and last dates of the input data.

Details

Upon import, the package uses `zoo` and `lubridate` to process the input date and temperature data. It reads in daily data with the time vector specified as either POSIXct or Date (e.g. "1982-01-01 02:00:00" or "1982-01-01"). The data may be an irregular time series, but date must be ordered. The function constructs a complete time series from the start date to the end date, and fills in the regions in the time series where temperature data are missing with NAs in the temperature vector. There must only be one temperature value per day otherwise the function takes the mean of multiple values. It is up to the user to calculate daily data from sub-daily measurements. Leap years are automatically accommodated by this function.

This function can handle some of missing days, but this is not a licence to actually use these data for the detection of anomalous thermal events. Hobday et al. (2016) recommend gaps of no more than 3 days, which may be adjusted by setting the maxPadLength argument of the ts2clm function. The longer and more frequent the gaps become the lower the fidelity of the annual climatology and threshold that can be calculated, which will not only have repercussions for the accuracy at which the event metrics can be determined, but also for the number of events that can be detected.

It is recommended that a climatology period of at least 30 years is specified in order to capture any decadal thermal periodicities.

A faster version of this function called make_whole_fast is enabled by default in ts2clm, and we assume (hopefully correctly) that the user is certain that her data do not have missing rows (dates) or duplicated values. This slower but more robust function (i.e. make_whole) may be selected as an argument to ts2clm in case the data may have some issues.