Plotly helpers

This S3 method for GeomFlame allows it to be implemented with plotly::ggplotly()

# S3 method for GeomFlame
geom2trace(data, params, p)

Arguments

data
This is a data.frame of information passed to this function from plotly:::layers2traces

params
This is a packet of specific information also passed to this functions from plotly:::layers2traces

p
This is the base plot created by calling ggplot, but is still passed to this functions from plotly:::layers2traces

Examples

ts_res <- ts2clm(data = sst_WA, climatologyPeriod = c("1982-01-01", "2011-12-31"))
ts_res_sub <- ts_res[10500:10800,]

library(ggplot2)
library(plotly)

#> 
#> Attaching package: ‘plotly’

#> The following object is masked from ‘package:ggplot2’:
#> 
#>     last_plot

#> The following object is masked from ‘package:stats’:
#> 
#>     filter

#> The following object is masked from ‘package:graphics’:
#> 
#>     layout

p <- ggplot(data = ts_res_sub, aes(x = t, y = temp)) +
  geom_flame(aes(y2 = thresh), fill = "salmon") +
  geom_line(aes(y = temp)) +
  geom_line(aes(y = seas), colour = "green") +
  geom_line(aes(y = thresh), colour = "red") +
  labs(x = "", y = "Temperature (°C)")
 plotly::ggplotly(p)