Calculate the categories of events.

Calculates the categories of a series of events as produced by detect in accordance with the naming scheme proposed in Hobday et al. (in review).

category(data, y = "temp", name = "Event", hemisphere = "South")

Arguments

data
The function receives the full (list) output from the detect function.

y
This is a column containing the measurement variable. If the column name differs from the default (i.e. temp), specify the name here.

name
If a value is provide here it will be used to name the events in the event_name column (see below) of the output. Default is "Event".

hemisphere
This argument informs the function within which hemisphere ("North" or "South") the data were collected so that it may correctly output the season column (see below). The default is "South".

Value

The function will return a tibble with results similar to those seen in Table 2 of Hobday et al. (in review). This provides the information necessary to appraise the extent of the events in the output of detect based on the category ranking scale. The category thresholds are calculated based on the difference between the seasonal climatology and threshold climatology produced by detect. The four category levels are then the difference multiplied by the category level.

The categories are:

Moderate

Events that have been detected, but with a maximum intensity that does not double the distance between the seasonal climatology and the threshold value.

Strong

Events with a maximum intensity that doubles the distance from the seasonal climatology and the threshold, but do not triple it.

Severe

Events that triple the aforementioned distance, but do not quadruple it.

Extreme

Events with a maximum intensity that is four times or greater the aforementioned distance. Scary stuff...

The definitions for the output columns are as follows:
event_no

The number of the event as determined by detect for reference between the outputs.

event_name

The name of the event. Generated from the name value provided and the year of the peak_date (see following) of the event. If no name value is provided the default ``Event'' is used. As proposed in Hobday et al. (in review), Moderate events are not given a name so as to prevent multiple repeat names within the same year. If two or more events ranked greater than Moderate are reported withiin the same year, they will be differentiated with the addition of a trailing letter (e.g. Event 2001 a, Event 2001 b).

peak_date

The date (day) on which the maximum intensity of the event was recorded.

category

The maximum category threshold reached/exceeded by the event.

i_max

The maximum intensity of the event above the threshold value.

duration

The total duration (days) of the event. Note that this includes any possible days when the measurement value y) may have dropped below the threshold value. Therefore, the proportion of the event duration (days) spent above certain thresholds may not add up to 100% (see following four items).

p_moderate

The proportion of the total duration (days) spent above at or the first threshold, but below any further thresholds.

p_strong

The proportion of the total duration (days) spent at or above the second threshold, but below any further thresholds.

p_severe

The proportion of the total duration (days) spent at or above the third threshold, but below the fourth threshold.

p_extreme

The proportion of the total duration (days) spent at or above the fourth and final threshold. There is currently no recorded event that has exceeded a hypothetical fifth threshold so none is calculated... yet..

season

The season(S) during which the event occurred. If the event occurred across two seasons this will be displayed as ``Winter/Spring''. Across three seasons as ``Winter-Summer''. Events lasting across four or more seasons are listed as ``Year-round''. December (June) is used here as the start of Austral (Boreal) summer.

References

Hobday et al. (in review). Categorizing and Naming Marine Heatwaves. Oceanography.

Examples

ts_dat <- make_whole(sst_WA)
res <- detect(ts_dat, climatology_start = "1983-01-01",
                         climatology_end = "2012-12-31")
res_cat <- category(res)
head(res_cat)

#> # A tibble: 6 x 11
#>   event_no event_name peak_date  category   i_max duration p_moderate p_strong
#>      <int> <fct>      <date>     <chr>      <dbl>    <dbl>      <dbl>    <dbl>
#> 1        1 Event 1984 1984-06-05 IV Extreme  1.92        5          0        0
#> 2        2 Event 1984 1984-06-19 IV Extreme  2.11        6          0        0
#> 3        3 Event 1984 1984-07-10 IV Extreme  2.22       19          0        0
#> 4        4 Event 1984 1984-10-23 IV Extreme  1.83        7          0        0
#> 5        5 Event 1984 1984-10-30 IV Extreme  1.49        5          0        0
#> 6        6 Event 1984 1984-11-14 IV Extreme  1.43        5          0        0
#> # ... with 3 more variables: p_severe <dbl>, p_extreme <dbl>, season <chr>