Calculate the categories of events.

Calculates the categories of a series of events as produced by detect_event in accordance with the naming scheme proposed in Hobday et al. (in review).

category(data, y = temp, S = TRUE, name = "Event")

Arguments

data
The function receives the full (list) output from the detect_event function.

y
This is a column containing the measurement variable. If the column name differs from the default (i.e. temp), specify the name here.

S
This argument informs the function if the data were collected in the southern hemisphere (TRUE, default) or the northern hemisphere (FALSE) so that it may correctly output the season column (see below).

name
If a character string (e.g. "Bohai Sea") is provide here it will be used to name the events in the event_name column (see below) of the output. If no value is provided the default output is "Event".

Value

The function will return a tibble with results similar to those seen in Table 2 of Hobday et al. (in review). This provides the information necessary to appraise the extent of the events in the output of detect_event based on the category ranking scale. The category thresholds are calculated based on the difference between the seasonal climatology and threshold climatology produced by detect_event. The four category levels are then the difference multiplied by the category level.

The definitions for the output columns are as follows:

event_no

The number of the event as determined by detect_event for reference between the outputs.

event_name

The name of the event. Generated from the name value provided and the year of the peak_date (see following) of the event. If no name value is provided the default "Event" is used. As proposed in Hobday et al. (in review), Moderate events are not given a name so as to prevent multiple repeat names within the same year. If two or more events ranked greater than Moderate are reported withiin the same year, they will be differentiated with the addition of a trailing letter (e.g. Event 2001 a, Event 2001 b). (still in development)

peak_date

The date (day) on which the maximum intensity of the event was recorded.

category

The maximum category threshold reached/exceeded by the event.

i_max

The maximum intensity of the event above the threshold value.

duration

The total duration (days) of the event. Note that this includes any possible days when the measurement value y) may have dropped below the threshold value. Therefore, the proportion of the event duration (days) spent above certain thresholds may not add up to 100% (see following four items).

p_moderate

The proportion of the total duration (days) spent above at or the first threshold, but below any further thresholds.

p_strong

The proportion of the total duration (days) spent at or above the second threshold, but below any further thresholds.

p_severe

The proportion of the total duration (days) spent at or above the third threshold, but below the fourth threshold.

p_extreme

The proportion of the total duration (days) spent at or above the fourth and final threshold. There is currently no recorded event that has exceeded a hypothetical fifth threshold so none is calculated... yet..

season

The season(S) during which the event occurred. If the event occurred across two seasons this will be displayed as "Winter/Spring". Across three seasons as "Winter-Summer". Events lasting across four or more seasons are listed as "Year-round". December (June) is used here as the start of Austral (Boreal) summer.

Details

An explanation for the categories is as follows:

ModerateEvents that have been detected, but with a maximum intensity that does not double the distance between the seasonal climatology and the threshold value.

StrongEvents with a maximum intensity that doubles the distance from the seasonal climatology and the threshold, but do not triple it.

SevereEvents that triple the aforementioned distance, but do not quadruple it.

ExtremeEvents with a maximum intensity that is four times or greater the aforementioned distance. Scary stuff...

References

Hobday et al. (in review). Categorizing and Naming Marine Heatwaves. Oceanography.

Examples

res_WA <- detect_event(ts2clm(sst_WA,
                       climatologyPeriod = c("1983-01-01", "2012-12-31")))
# Note that the name argument expects a character vector
cat_WA <- category(res_WA, name = "WA")
tail(cat_WA)

#> # A tibble: 6 x 11
#>   event_no event_name peak_date  category   i_max duration p_moderate p_strong
#>      <int> <fct>      <date>     <chr>      <dbl>    <int>      <dbl>    <dbl>
#> 1       38 WA 2010    2010-09-29 II Strong   2.58       18         78       22
#> 2       40 WA 2010    2010-12-02 II Strong   2.66       15         73       27
#> 3       31 WA 2008    2008-04-14 III Severe  3.77       34         62       29
#> 4       22 WA 1999    1999-05-22 II Strong   3.6        95         59       41
#> 5       58 WA 2014    2014-08-31 II Strong   2.2        12         58       42
#> 6       42 WA 2011    2011-02-28 IV Extreme  6.51       60         47       32
#> # ... with 3 more variables: p_severe <dbl>, p_extreme <dbl>, season <chr>

# If the data were collected in the northern hemisphere
# we must let the funciton know this as seen below
res_Med <- detect_event(ts2clm(sst_Med,
                        climatologyPeriod = c("1983-01-01", "2012-12-31")))
cat_Med <- category(res_Med, S = FALSE, name = "Med")
tail(cat_Med)

#> # A tibble: 6 x 11
#>   event_no event_name peak_date  category   i_max duration p_moderate p_strong
#>      <int> <fct>      <date>     <chr>      <dbl>    <int>      <dbl>    <dbl>
#> 1       55 Med 2011   2011-05-25 II Strong   4.05        8         62       38
#> 2       23 Med 1999   1999-04-06 II Strong   1.88        5         60       40
#> 3       28 Med 2003   2003-06-20 II Strong   5.02       30         57       43
#> 4       62 Med 2012   2012-08-20 II Strong   4.22       18         56       44
#> 5       44 Med 2007   2007-04-25 III Severe  4.02       19         42       53
#> 6       70 Med 2014   2014-10-18 II Strong   3.31      110         35       65
#> # ... with 3 more variables: p_severe <dbl>, p_extreme <dbl>, season <chr>